Reduce scenario
