+++ features/product/managing_product_variants/checking_products_in_the_shop_while_editing_their_product_variants.feature
@@ -1,0 +1,26 @@
+@managing_product_variants
+Feature: Checking products in the shop while editing their product variants
+In order to check a product in shop in all channels it is available in
+    As an Administrator
+    I want to choose a channel in which I will be viewing it
+
+    Background:
+        Given the store operates on a channel named "United States" with hostname "goodcars.com"
+        And the store also operates on a channel named "Europe" with hostname "goodcars.eu"
+        And the store has a product "Bugatti" priced at "$200000.00" in "United States" channel
+        And I am logged in as an administrator
+
+    @ui
+    Scenario: Accessing product show page in shop from the product variant edit page where product is available in more than one channel
+        Given this product is also available in the "Europe" channel
+        And this product has "Red" variant priced at "$220000.00" in "Europe" channel
+        And I want to modify the "Bugatti" product variant
+        And I choose to show this product in the "Europe" channel
+        Then I should see this product in the "Europe" channel in the shop
+
+    @ui
+    Scenario: Accessing product show page in shop from the product variant edit page where product is available in one channel
+        Given this product has "Red" variant priced at "$220000.00" in "United States" channel
+        And I want to modify the "Bugatti" product variant
+        And I choose to show this product in this channel
+        Then I should see this product in the "United States" channel in the shop
