+++ src/Sylius/Behat/Context/Setup/ProductContext.php
@@ -287,7 +287,6 @@ final class ProductContext implements Context
 
     /**
      * @Given /^the (product "[^"]+") has(?:| a) "([^"]+)" variant priced at ("[^"]+")$/
-     * @Given /^the (product "[^"]+") has(?:| a) "([^"]+)" variant priced at ("[^"]+") in ("([^"]+)" channel)$/
      * @Given /^(this product) has "([^"]+)" variant priced at ("[^"]+")$/
      * @Given /^(this product) has "([^"]+)" variant priced at ("[^"]+") in ("([^"]+)" channel)$/
      */
