+++ src/Sylius/Behat/Context/Ui/Admin/ManagingProductsContext.php
@@ -27,7 +27,6 @@ use Sylius\Behat\Page\Admin\Product\UpdateSimpleProductPageInterface;
 use Sylius\Behat\Page\Admin\ProductReview\IndexPageInterface as ProductReviewIndexPageInterface;
 use Sylius\Behat\Page\Admin\ProductVariant\CreatePageInterface as VariantCreatePageInterface;
 use Sylius\Behat\Page\Admin\ProductVariant\GeneratePageInterface;
-use Sylius\Behat\Page\Shop\Product\ShowPage;
 use Sylius\Behat\Service\NotificationCheckerInterface;
 use Sylius\Behat\Service\Resolver\CurrentPageResolverInterface;
 use Sylius\Behat\Service\SharedStorageInterface;
@@ -50,9 +49,6 @@ final class ManagingProductsContext implements Context
 
     /** @var IndexPageInterface */
     private $indexPage;
-
-    /** @var ShowPage */
-    private $showPage;
 
     /** @var UpdateSimpleProductPageInterface */
     private $updateSimpleProductPage;
@@ -83,7 +79,6 @@ final class ManagingProductsContext implements Context
         CreateSimpleProductPageInterface $createSimpleProductPage,
         CreateConfigurableProductPageInterface $createConfigurableProductPage,
         IndexPageInterface $indexPage,
-        ShowPage $showPage,
         UpdateSimpleProductPageInterface $updateSimpleProductPage,
         UpdateConfigurableProductPageInterface $updateConfigurableProductPage,
         ProductReviewIndexPageInterface $productReviewIndexPage,
@@ -97,7 +92,6 @@ final class ManagingProductsContext implements Context
         $this->createSimpleProductPage = $createSimpleProductPage;
         $this->createConfigurableProductPage = $createConfigurableProductPage;
         $this->indexPage = $indexPage;
-        $this->showPage = $showPage;
         $this->updateSimpleProductPage = $updateSimpleProductPage;
         $this->updateConfigurableProductPage = $updateConfigurableProductPage;
         $this->productReviewIndexPage = $productReviewIndexPage;
@@ -122,14 +116,6 @@ final class ManagingProductsContext implements Context
     public function iWantToCreateANewConfigurableProduct()
     {
         $this->createConfigurableProductPage->open();
-    }
-
-    /**
-     * @When I access :product product edit page
-     */
-    public function iAccessProductEditPage(ProductInterface $product): void
-    {
-        $this->indexPage->showProductEditPage($product->getName());
     }
 
     /**
@@ -234,19 +220,19 @@ final class ManagingProductsContext implements Context
     }
 
     /**
-     * @When I show this product in the :channel channel
+     * @When I choose to show this product in the :channel channel
      */
-    public function iShowThisProductInTheChannel(string $channel): void
+    public function iChooseToShowThisProductInTheChannel(string $channel): void
     {
-        $this->resolveCurrentPage()->showProductInChannel($channel);
+        $this->updateSimpleProductPage->showProductInChannel($channel);
     }
 
     /**
-     * @When I show this product in this channel
+     * @When I choose to show this product in this channel
      */
-    public function iShowThisProductInThisChannel(): void
+    public function iChooseToShowThisProductInThisChannel(): void
     {
-        $this->resolveCurrentPage()->showProductInSingleChannel();
+        $this->updateSimpleProductPage->showProductInSingleChannel();
     }
 
     /**
@@ -1023,15 +1009,6 @@ final class ManagingProductsContext implements Context
     public function iShouldSeeInventoryOfThisProduct(): void
     {
         Assert::true($this->updateSimpleProductPage->hasInventoryTab());
-    }
-
-    /**
-     * @Then /^I should see (this product) in the ("([^"]*)" channel) in shop$/
-     */
-    public function iShouldSeeThisProductInTheChannelInShop(ProductInterface $product, ChannelInterface $channel): void
-    {
-        Assert::true(null !== strpos($this->showPage->getCurrentUrl(), $channel->getHostname()));
-        Assert::same($this->showPage->getName(), $product->getName());
     }
 
     /**
