+++ src/Sylius/Behat/Page/Admin/Product/IndexPage.php
@@ -37,11 +37,6 @@ final class IndexPage extends CrudIndexPage implements IndexPageInterface
         $this->imageExistenceChecker = $imageExistenceChecker;
     }
 
-    public function goToVariantList(): void
-    {
-        $this->getDocument()->clickLink('List variants');
-    }
-
     public function filterByTaxon(string $taxonName): void
     {
         $this->getElement('taxon_filter', ['%taxon%' => $taxonName])->click();
@@ -53,11 +48,6 @@ final class IndexPage extends CrudIndexPage implements IndexPageInterface
         $imageUrl = $productRow->find('css', 'img')->getAttribute('src');
 
         return $this->imageExistenceChecker->doesImageWithUrlExist($imageUrl, 'sylius_admin_product_thumbnail');
-    }
-
-    public function showProductEditPage(string $productName): void
-    {
-        $this->getElement('editButton', ['%productName%' => $productName])->clickLink('Edit');
     }
 
     public function showProductPage(string $productName): void
@@ -72,7 +62,6 @@ final class IndexPage extends CrudIndexPage implements IndexPageInterface
     protected function getDefinedElements(): array
     {
         return array_merge(parent::getDefinedElements(), [
-            'editButton' => '.table tr:contains("%productName%") td:nth-child(7)',
             'taxon_filter' => '.sylius-tree__item a:contains("%taxon%")',
         ]);
     }
