+++ src/Sylius/Behat/Resources/config/services/contexts/ui.xml
@@ -151,7 +151,6 @@
             <argument type="service" id="sylius.behat.page.admin.product.create_simple" />
             <argument type="service" id="sylius.behat.page.admin.product.create_configurable" />
             <argument type="service" id="sylius.behat.page.admin.product.index" />
-            <argument type="service" id="sylius.behat.page.shop.product.show" />
             <argument type="service" id="sylius.behat.page.admin.product.update_simple" />
             <argument type="service" id="sylius.behat.page.admin.product.update_configurable" />
             <argument type="service" id="sylius.behat.page.admin.product_review.index" />
@@ -195,7 +194,6 @@
             <argument type="service" id="sylius.behat.page.admin.product_variant.create" />
             <argument type="service" id="sylius.behat.page.admin.product_variant.index" />
             <argument type="service" id="sylius.behat.page.admin.product_variant.update" />
-            <argument type="service" id="sylius.behat.page.shop.product.show" />
             <argument type="service" id="sylius.behat.page.admin.product_variant.generate" />
             <argument type="service" id="sylius.behat.current_page_resolver" />
             <argument type="service" id="sylius.behat.notification_checker" />
@@ -203,7 +201,6 @@
 
         <service id="sylius.behat.context.ui.admin.browsing_product_variants" class="Sylius\Behat\Context\Ui\Admin\BrowsingProductVariantsContext">
             <argument type="service" id="sylius.behat.page.admin.product_variant.index" />
-            <argument type="service" id="sylius.behat.page.admin.product.index" />
             <argument type="service" id="sylius.product_variant_resolver.default" />
         </service>
 
@@ -439,6 +436,10 @@
             <argument type="service" id="sylius.behat.page.shop.product.index" />
             <argument type="service" id="sylius.behat.page.shop.product_reviews.index" />
             <argument type="service" id="sylius.behat.page.error" />
+        </service>
+
+        <service id="sylius.behat.context.ui.shop.browsing_product" class="Sylius\Behat\Context\Ui\Shop\BrowsingProductContext">
+            <argument type="service" id="sylius.behat.page.shop.product.show" />
         </service>
 
         <service id="sylius.behat.context.ui.shop.product_review" class="Sylius\Behat\Context\Ui\Shop\ProductReviewContext">
