+++ src/Sylius/Behat/Resources/config/suites/ui/product/managing_product_variants.yml
@@ -36,6 +36,7 @@ default:
                 - sylius.behat.context.ui.admin.browsing_product_variants
                 - sylius.behat.context.ui.admin.managing_product_variants
                 - sylius.behat.context.ui.admin.notification
+                - sylius.behat.context.ui.shop.browsing_product
 
             filters:
                 tags: "@managing_product_variants && @ui"
