+++ src/Sylius/Behat/Resources/config/suites/ui/product/managing_products.yml
@@ -45,6 +45,7 @@ default:
                 - sylius.behat.context.ui.admin.browsing_product_variants
                 - sylius.behat.context.ui.admin.managing_products
                 - sylius.behat.context.ui.admin.notification
+                - sylius.behat.context.ui.shop.browsing_product
                 - sylius.behat.context.ui.shop.locale
 
             filters:
