+++ src/Sylius/Bundle/AdminBundle/Resources/views/Product/Show/_header.html.twig
@@ -9,5 +9,5 @@
         </h1>
         {% include "@SyliusAdmin/Product/Show/_breadcrumb.html.twig" %}
     </div>
-    {% include '@SyliusAdmin/Product/shopProductShowPagePerChannelButton.html.twig' with {'product': product} %}
+    {% include '@SyliusAdmin/Product/_showInShopButton.html.twig' %}
 </div>
