+++ src/Sylius/Bundle/AdminBundle/Resources/views/Product/_showInShopButton.html.twig
@@ -1,14 +1,15 @@
-{% if product.channels|length > 1 and product.channels|filter(channel => channel.enabled == true) %}
+{% set enabledChannels = product.channels|filter(channel => channel.enabled == true) %}
+{% if enabledChannels|length > 1 %}
     <div class="middle aligned column">
         <div class="ui right floated buttons">
             <div class="ui floating dropdown labeled icon button">
                 <i class="share alternate icon"></i>
                 <span class="text">
-                {{ 'sylius.ui.show_product_in_shop_page'|trans }}
+                    {{ 'sylius.ui.show_product_in_shop_page'|trans }}
                 </span>
                 <div class="menu">
                     <div class="scrolling menu">
-                        {% for channel in product.channels if channel.enabled == true %}
+                        {% for channel in enabledChannels %}
                             {% set url = channel.hostname is not null ? 'http://' ~ channel.hostname ~ path('sylius_shop_product_show', {'slug': product.slug, '_locale': channel.defaultLocale.code}) : url('sylius_shop_product_show', {'slug': product.slug, '_locale': channel.defaultLocale.code}) %}
                             <a href="{{ url|raw }}" class="item">
                                 <i class="angle right icon"></i>
@@ -21,10 +22,10 @@
             </div>
         </div>
     </div>
-{% elseif product.channels|length == 1 %}
+{% elseif enabledChannels|length == 1 %}
     <div class="middle aligned column">
         <div class="ui right floated buttons">
-            {% for channel in product.channels if channel.enabled == true %}
+            {% for channel in enabledChannels %}
                 {% set url = channel.hostname is not null ? 'http://' ~ channel.hostname ~ path('sylius_shop_product_show', {'slug': product.slug, '_locale': channel.defaultLocale.code}) : url('sylius_shop_product_show', {'slug': product.slug, '_locale': channel.defaultLocale.code}) %}
                 <a class="ui labeled icon button" href="{{ url|raw }}">
                     <i class="angle right icon"></i>
