+++ src/Sylius/Bundle/AdminBundle/Resources/views/ProductVariant/Update/_toolbar.html.twig
@@ -1,1 +1,1 @@
-{% include '@SyliusAdmin/Product/shopProductShowPagePerChannelButton.html.twig' with {'product': product_variant.product} %}
+{% include '@SyliusAdmin/Product/_showInShopButton.html.twig' with {'product': product_variant.product} %}
